added ARP + int64 is now really 64 bit, TOOO: fix mising frame error

+++ include/common/types.h
@@ -22,8 +22,8 @@ namespace coolOS{
         typedef unsigned short uint16_t;
         typedef int int32_t;
         typedef unsigned int uint32_t;
-        typedef long int64_t;
-        typedef unsigned long uint64_t;
+        typedef long long int64_t;
+        typedef unsigned long long uint64_t;
         
         typedef const char* string;
         typedef uint32_t size_t;
